Fix Curse trying to loop through creatures with no effects Fix telekenisis not grabbing the proper player handler.

+++ src/com/herocraftonline/dev/heroes/skill/skills/SkillSmoke.java
@@ -58,7 +58,7 @@ public class SkillSmoke extends ActiveSkill {
     public boolean use(Hero hero, String[] args) {
         broadcastExecuteText(hero);
 
-        int duration = getSetting(hero.getHeroClass(), "duration", 20000);
+        long duration = getSetting(hero.getHeroClass(), "duration", 20000);
         Player player = hero.getPlayer();
         player.getWorld().playEffect(player.getLocation(), org.bukkit.Effect.SMOKE, 4);
         hero.addEffect(new SmokeEffect(this, duration));
